Derive buffer name from working directory of launched process by default.
This is required to avoid a mismatch between a terminal buffer's name
and its actual current working directory, which might differ from the
previously used `default-directory'.

An example of such a mismatch would be a multi-term that was advised by
projectile to launch in the current project's root directory rather than
in `default-directory`.

Note that this functionality depends on the `lsof` command (available on
both linux and osx platforms).